[Bronze V] Title: 팩토리얼 2, Time: 0 ms, Memory: 1112 KB -BaekjoonHub

+++ C++11/백준/Bronze/27433. 팩토리얼 2/팩토리얼 2.cc
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main(){
-    long a; long ans=1L;
-    scanf("%ld", &a);
+    int a; long ans=1;
+    scanf("%d", &a);
     for(int i =0; i<a; ++i){
         ans*=i+1;
     }
